Changed valid js extensions to jsx
