+++ component/home-page/HomeWithBackground.jsx
