--- pages/index.js
@@ -1,11 +1,0 @@
-import HomeWithBackground from '../component/home-page/HomeWithBackground';
-
-export default function Home() {
-
-   return (
-     <main>
-      <HomeWithBackground/> {/* Render the HomeWithBackground component. */}
-     </main>
-   );
-   
- }
