+++ pages/index.jsx
@@ -1,0 +1,9 @@
+import HomeWithBackground from '@/component/home-page/HomeWithBackground';
+
+export default function Home() {
+  return (
+    <main>
+      <HomeWithBackground /> {/* Render the HomeWithBackground component. */}
+    </main>
+  );
+}
